Refactor pacckage containing Activator classes. Change how the Config static field is created from the Activator
