+++ distro/karaf/bundles/gateway-osgi-servlet/pom.xml
@@ -98,7 +98,7 @@
                 <configuration>
                     <instructions>
                         <Bundle-SymbolicName>${project.groupId}.${project.artifactId}</Bundle-SymbolicName>
-                        <Bundle-Activator>io.apiman.gateway.platforms.osgi.Activator</Bundle-Activator>
+                        <Bundle-Activator>io.apiman.gateway.osgi.servlet.Activator</Bundle-Activator>
                         <Export-Package>
                             io.apiman.gateway.platforms.war,
                             io.apiman.gateway.platforms.war.filters,
@@ -106,6 +106,11 @@
                             io.apiman.gateway.platforms.war.listeners,
                             io.apiman.gateway.platforms.war.servlets
                         </Export-Package>
+                        <Import-Package>!
+                            !io.apiman.gateway.osgi.servlet,
+                            !io.apiman.gateway.platforms.war*,
+                            *
+                        </Import-Package>
                         <Embed-Dependency>
                             *;scope=compile;artifactId=!junit|mockito-all
                         </Embed-Dependency>
