+++ distro/karaf/bundles/gateway-osgi/src/main/java/io/apiman/gateway/osgi/Activator.java
@@ -1,4 +1,4 @@
-package io.apiman.gateway.platforms.osgi;
+package io.apiman.gateway.osgi;
 
 import io.apiman.common.servlet.ApimanCorsFilter;
 import io.apiman.common.servlet.DisableCachingFilter;
